backend/pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["rubanaa <andreygrigorev0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "*"
djangorestframework = "*"
django-cors-headers = "*"
djangorestframework-simplejwt = "*"
djangorestframework-camel-case  = "*"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
